A very rough first cut of the docbook build scripts. This does a few things:
1. Builds the pdf and html docs when you execute "mvn site"
2. Obliterates the ugly and annoying default maven site behaviour
3. Applies style sheets, css, etc
4. Sets up some reasonable defaults for TOC and other styling

What isn't done yet is pulling across more than just a few pages of sample docbook. This will take lots of effort
And of course, making it more attractive, which can wait until later.

I've put in some sample docbook elements which are interesting: lists, sections, link and xref, code snippets.

This module has not yet been tied into the main Cayenne pom and it is in a folder location designed to keep out of the way of the existing documentation. When it is complete, we can move it into a better location.



git-svn-id: https://svn.apache.org/repos/asf/cayenne/main/trunk@1098308 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 951055102fc59c52a2c4c585dd6d3644142c930d