Bugfix: No methods for queries with qualifier parameters generated.
Cgen should generate a perform-method for queries with qualifiers (e.g. "username = $ausername") stored in the datamap, but there is no method generated. Queries without qualifiers get a method.

If the alias map for Entity.lastPathComponent is null then an exception is thrown later in org.apache.cayenne.map.PathComponentIterator at the line "this.aliasMap = Objects.requireNonNull(aliasMap)". This means the qualifier is seen as invalid and therefore no perform-method is generated.


Former-commit-id: 05ec056740646fdee69d15b3fc8c1e665794f9fd